+++ common/src/main/java/com/cozary/colored_water/mixin/LiquidBlockRendererMixin.java
@@ -7,6 +7,7 @@ import com.mojang.blaze3d.vertex.VertexConsumer;
 import net.minecraft.client.renderer.block.LiquidBlockRenderer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockAndTintGetter;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
@@ -30,7 +31,9 @@ public class LiquidBlockRendererMixin {
         require = 0
     )
     private FluidState coloredWater$modifyWaterloggedFluidState(FluidState fluidState, BlockAndTintGetter level, BlockPos pos, VertexConsumer buffer, BlockState blockState) {
-        if (blockState != null && blockState.hasProperty(BlockStateProperties.WATERLOGGED) && blockState.getValue(BlockStateProperties.WATERLOGGED)) {
+        boolean isWaterlogged = blockState != null && blockState.hasProperty(BlockStateProperties.WATERLOGGED) && blockState.getValue(BlockStateProperties.WATERLOGGED);
+        boolean isBubbleColumn = blockState != null && blockState.is(Blocks.BUBBLE_COLUMN);
+        if (isWaterlogged || isBubbleColumn) {
             if (level != null && pos != null) {
                 BlockEntity be = level.getBlockEntity(pos);
                 if (be instanceof ColoredWaterBlockEntity coloredBe) {